Move the handling of events to a deque and not waiting on socket
In the past in on_start() we would just loop on the socket. This
worked, but was really bad practice as the socket would just fill with
events while a behavior was running. This puts the events in a local
queue (where they should be) and lets the socket stay clean. This
should also let us implement such things as an emergency stop or
filter out duplicate events.

Right now though, the effect should remain the same.